Add direct buy and rental checkout flows

+++ frontend/app/src/main/java/com/example/my_mobile_app/api/dto/CreateRentalRequest.java
@@ -7,7 +7,7 @@ public class CreateRentalRequest {
     public String shippingAddress;
     /** COD | MoMo. */
     public String paymentMethod;
-    public Double shippingFee;
+    public Long shippingFee;
 
     public CreateRentalRequest() {}
